moves coin animation to tools component & starts animation on socket emit :sunglasses:
